🎨 Remove unneeded conditions, remove code repeats, add new response type

+++ server/src/authorized_routers.py
@@ -1,9 +1,13 @@
 from datetime import datetime, timedelta
 from fastapi import APIRouter, Depends
-from cache import Cache
 from database import Database
-from dependencies import get_database, get_cache
-from responses import SUCCESS_INSERTION, SUCCESS_DELETION, serialize_instances
+from dependencies import get_database, get_request_data_from_cache
+from responses import (
+    SUCCESS_INSERTION,
+    SUCCESS_DELETION,
+    ODOO_INSTANCE_NOT_EXIST,
+    serialize_instances
+)
 
 
 router = APIRouter(prefix="/authorized")
@@ -11,10 +15,8 @@ router = APIRouter(prefix="/authorized")
 
 @router.get("/post_instance")
 async def auth_post_instance(
-        uuid: str,
-        cache: Cache = Depends(get_cache),
+        request_data: dict = Depends(get_request_data_from_cache),
         db: Database = Depends(get_database)):
-    request_data = await cache.get_record(uuid)
     if not await db.user_exists(request_data["access_token"]):
         due_time = timedelta(seconds=int(request_data["expires_in"]))
         due_date = datetime.now() + due_time
@@ -23,15 +25,11 @@ async def auth_post_instance(
             request_data["refresh_token"],
             due_date.date()
         )
-    if db.odoo_instance_exists(
-            request_data["access_token"],
-            request_data["url"],
-            request_data["db_name"]):
-        await db.delete_odoo_instance(
-            request_data["access_token"],
-            request_data["url"],
-            request_data["db_name"]
-        )
+    await db.delete_odoo_instance(
+        request_data["access_token"],
+        request_data["url"],
+        request_data["db_name"]
+    )
     await db.insert_odoo_instance(
         request_data["access_token"],
         request_data["url"],
@@ -44,23 +42,25 @@ async def auth_post_instance(
 
 @router.get("/delete_instance")
 async def auth_delete_instance(
-        uuid: str,
-        cache: Cache = Depends(get_cache),
+        request_data: dict = Depends(get_request_data_from_cache),
         db: Database = Depends(get_database)):
-    request_data = await cache.get_record(uuid)
-    await db.delete_odoo_instance(
-        request_data["access_token"],
-        request_data["url"],
-        request_data["db_name"]
-    )
-    return SUCCESS_DELETION
+    if await db.odoo_instance_exists(
+            request_data["access_token"],
+            request_data["url"],
+            request_data["db_name"]):
+        await db.delete_odoo_instance(
+            request_data["access_token"],
+            request_data["url"],
+            request_data["db_name"]
+        )
+        return SUCCESS_DELETION
+    else:
+        return ODOO_INSTANCE_NOT_EXIST
 
 
 @router.get("/get_instance")
 async def auth_get_instance(
-        uuid: str,
-        cache: Cache = Depends(get_cache),
+        request_data: dict = Depends(get_request_data_from_cache),
         db: Database = Depends(get_database)):
-    request_data = await cache.get_record(uuid)
     instances = await db.get_instances_of_user(request_data["access_token"])
     return serialize_instances(instances)
